% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOboot.one.R
\name{AOboot_one}
\alias{AOboot_one}
\alias{print.AOboot_one}
\title{AOboot.one Class}
\usage{
AOboot_one(type.aov, factor, anova, em, no.test, ph)
}
\arguments{
\item{type.aov}{Character string giving the type of ANOVA computed.}

\item{factor}{Names of groups in the entered factor.}

\item{anova}{Results of the ANOVA.}

\item{em}{Bootstrapped estimated marginal means.}

\item{no.test}{Number of tests conducted that did not produce errors.}

\item{ph}{Bootstrapped post hoc tests.}
}
\value{
An object of class \code{"AOboot.one"}.
}
\description{
A S3 class to represent one-way ANOVAs.
}
