% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOboot.two.R
\name{AOboot_two}
\alias{AOboot_two}
\alias{print.AOboot_two}
\title{AOboot.two Class}
\usage{
AOboot_two(
  type.aov,
  factor1,
  factor2,
  anova,
  em.1,
  no.test1,
  ph.1,
  em.2,
  no.test2,
  ph.2,
  em.3,
  no.test3,
  ph.3,
  em.4,
  no.test4,
  ph.4
)
}
\arguments{
\item{type.aov}{Character string giving the type of ANOVA computed.}

\item{factor1}{Names of groups in the first factor.}

\item{factor2}{Names of groups in the second factor.}

\item{anova}{Results of the ANOVA.}

\item{em.1}{Bootstrapped estimated marginal means for factor 1.}

\item{no.test1}{Number of bootstrapped tests conducted for factor 1 that did
not produce errors.}

\item{ph.1}{Bootstrapped post hoc tests for factor 1.}

\item{em.2}{Bootstrapped estimated marginal means for factor 2.}

\item{no.test2}{Number of bootstrapped tests conducted for factor 2 that did
not produce errors.}

\item{ph.2}{Bootstrapped post hoc tests for factor 2.}

\item{em.3}{Bootstrapped estimated marginal means for factor 1 by factor 2.}

\item{no.test3}{Number of bootstrapped tests conducted for factor 1 by factor
2 that did not produce errors.}

\item{ph.3}{Bootstrapped post hoc tests for factor 1 by factor 2.}

\item{em.4}{Bootstrapped estimated marginal means for factor 2 by factor 1.}

\item{no.test4}{Number of bootstrapped tests conducted for factor 2 by factor
1 that did not produce errors.}

\item{ph.4}{Bootstrapped post hoc tests for factor 2 by factor 1.}
}
\value{
An object of class \code{"AOboot.two"}.
}
\description{
A S3 class to represent two-way ANOVAs.
}
